refactor: Change Category ID to Long and add OrderRepository
This commit introduces several improvements and fixes across the application:

- The primary key for the Category entity has been changed from Integer to Long to prevent potential overflow and align with best practices. This change is propagated to the `ManagerCategoryController` and `ManagerProductController`.

- The `mssql-jdbc` dependency version is now explicitly defined in `pom.xml` to ensure build consistency and prevent potential conflicts.

- A default status of "Active" is now set for new User entities, ensuring data integrity upon creation.

- A new `OrderRepository` has been implemented with methods to find orders by user, enabling future features related to order history.

+++ src/main/java/com/flowerShop1/controller/ManagerProductController.java
@@ -113,7 +113,7 @@ public class ManagerProductController {
     public String updateProduct(@PathVariable Integer productId,
                                 @Valid @ModelAttribute("product") Product product,
                                 BindingResult result,
-                                @RequestParam(value = "categoryId", required = false) Integer categoryId,
+                                @RequestParam(value = "categoryId", required = false) Long categoryId,
                                 @RequestParam(value = "supplierId", required = false) Integer supplierId,
                                 @RequestParam(value = "imageFile", required = false) MultipartFile imageFile,
                                 Model model,
@@ -165,7 +165,7 @@ public class ManagerProductController {
     public String createNewProduct(@Valid @ModelAttribute("product") Product product,
                                    BindingResult result,
                                    @RequestParam("imageFile") MultipartFile imageFile,
-                                   @RequestParam("categoryId") Integer categoryId,
+                                   @RequestParam("categoryId") Long categoryId,
                                    @RequestParam("supplierId") Integer supplierId,
                                    Model model,
                                    RedirectAttributes ra) throws IOException {
